+++ RELEASE_NOTES.md
@@ -7,7 +7,7 @@
 - Recover archived events on upgrade (3.4.0)
 - Improve upgrade experience (3.4.0)
 - Fixed application startup (3.4.1)
-- Detect antivirus software and suggest adding exclusions on _Sharly Chess_ folder (3.5.0)
+- Detect antivirus software and add Windows Defender exclusion on _Sharly Chess_ folder (3.5.0)
 
 ## Events